Messages BE updated to get, create, update and delete only logged in users

+++ src/server/api/messages.js
@@ -1,16 +1,9 @@
-/*
-GET /api/messages: Retrieve a list of all messages.
-GET /api/messages/:id: Retrieve a specific message by ID.
-POST /api/messages: Create a new message.
-PATCH /api/messages/:id: Update an existing message.
-DELETE /api/messages/:id: Delete a message by ID.
-*/
-
 const { ServerError } = require("../errors");
 const prisma = require("../prisma");
 
 const router = require("express").Router();
 
+// Gets all messages
 router.get("/", async (req, res, next) => {
   try {
     const messages = await prisma.message.findMany();
@@ -20,10 +13,13 @@ router.get("/", async (req, res, next) => {
   }
 });
 
-router.get("/user/:userId", async (req, res, next) => {
+// Gets all messages sent by logged in user
+router.get("/sent", async (req, res, next) => {
   try {
+    const senderId = +res.locals.user.id;
+
     const messages = await prisma.message.findMany({
-      where: { userId: +res.locals.user.id },
+      where: { senderId },
     });
 
     res.json(messages);
@@ -32,6 +28,56 @@ router.get("/user/:userId", async (req, res, next) => {
   }
 });
 
+// Gets all messages sent by logged in user
+router.get("/sent/:id", async (req, res, next) => {
+  try {
+    const messageId = +req.params.id;
+
+    const messages = await prisma.message.findUnique({
+      where: { id: messageId },
+    });
+
+    res.json(messages);
+  } catch (err) {
+    next(err);
+  }
+});
+
+// Gets all messages sent by logged in user
+router.get("/received", async (req, res, next) => {
+  try {
+    const recipientId = +res.locals.user.id;
+
+    const messages = await prisma.message.findMany({
+      where: { recipientId },
+    });
+
+    res.json(messages);
+  } catch (err) {
+    next(err);
+  }
+});
+
+// Gets all messages sent by logged in user
+router.get("/received/:id", async (req, res, next) => {
+  try {
+    const messageId = +req.params.id;
+
+    const message = await prisma.message.findUnique({
+      where: { id: messageId },
+    });
+
+    if (!message) {
+      return res.status(404).json({ error: "Message not found" });
+    }
+
+    res.json(message);
+  } catch (err) {
+    next(err);
+  }
+});
+
+// Creates a message from logged in user to me
 router.post("/", async (req, res, next) => {
   try {
     const { content } = req.body;
@@ -43,7 +89,11 @@ router.post("/", async (req, res, next) => {
     const fixedRecipientId = 21;
 
     const newMessage = await prisma.message.create({
-      data: { content, user: { connect: { id: fixedRecipientId } } },
+      data: {
+        content,
+        senderId: +res.locals.user.id,
+        recipientId: fixedRecipientId,
+      },
     });
 
     res.status(201).json(newMessage);
@@ -52,7 +102,8 @@ router.post("/", async (req, res, next) => {
   }
 });
 
-router.patch("/:id", async (req, res, next) => {
+// Updates a message sent by a logged in user
+router.patch("/sent/:id", async (req, res, next) => {
   try {
     const messageId = +req.params.id;
     const { content } = req.body;
@@ -69,9 +120,26 @@ router.patch("/:id", async (req, res, next) => {
   }
 });
 
-router.delete("/:id", async (req, res, next) => {
+router.delete("/sent/:id", async (req, res, next) => {
   try {
     const messageId = +req.params.id;
+    const userId = res.locals.user.id;
+
+    const message = await prisma.message.findUnique({
+      where: { id: messageId },
+    });
+
+    if (!message) {
+      return res.status(404).json({ error: "Message not found" });
+    }
+
+    if (message.senderId !== userId) {
+      return res.status(403).json({
+        error:
+          "This message does not belong to you!!! You cannot delete this message.",
+      });
+    }
+
     await prisma.message.delete({ where: { id: messageId } });
     res.sendStatus(204);
   } catch (err) {
